+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- All `always_ff` use `@(posedge clk or negedge rstn)`.
+
 ## [0.1.1] - 2026-08-10
 
 ### Added